Extract ContactsAppBar and ContactsList components
- Moved `ContactsAppBar` and `ContactsList` to separate composable functions for better organization.
- Added internet permission to AndroidManifest.
- Renamed and moved `CircularAvatar` and `ContactItem` to the components package.
